gf_item_carousel row cont double->int change

+++ lib/components/carousel/gf_items_carousel.dart
@@ -36,7 +36,7 @@ class GFItemsCarousel extends StatefulWidget {
   }) : super(key: key);
 
   /// Count of visible cells
-  final double rowCount;
+  final int rowCount;
 
   /// The widgets to be shown as sliders.
   final List<Widget> children;
